gen-ai-course/14_ai_projects/schedule_parser_calendar_sync/pyproject.toml
[project]
name = "schedule-parser-calendar-sync"
version = "0.1.0"
description = "Parse schedule screenshots via OCR and sync to Google Calendar"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "langchain-groq>=0.1.0",
    "langchain-core>=0.2.0",
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.2.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "google-cloud-vision>=3.5.0",
    "google-api-python-client>=2.100.0",
    "google-auth-oauthlib>=1.2.0",
    "Pillow>=10.0.0",
    "python-multipart>=0.0.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/schedule_parser"]

[tool.hatch.build.targets.sdist]
include = ["/src"]
